Fix build error, update dependencies
