+++ package.json
@@ -14,7 +14,7 @@
     "@carbon/elements": "^10.7.0",
     "@carbon/icons-react": "^10.6.1",
     "@reach/router": "^1.2.1",
-    "@react-google-maps/api": "^1.7.12",
+    "@react-google-maps/api": "^1.8.0",
     "add": "^2.0.6",
     "axios": "^0.19.0",
     "babel-plugin-styled-components": "^1.10.6",
@@ -24,21 +24,21 @@
     "color": "^3.1.2",
     "dotenv": "^8.2.0",
     "formik": "^2.0.6",
-    "gatsby": "^2.18.4",
+    "gatsby": "^2.18.7",
     "gatsby-awesome-pagination": "^0.3.5",
-    "gatsby-background-image": "^0.9.7",
+    "gatsby-background-image": "^0.9.11",
     "gatsby-image": "^2.2.34",
-    "gatsby-plugin-gtag": "^1.0.11",
+    "gatsby-plugin-gtag": "^1.0.12",
     "gatsby-plugin-mailchimp": "^5.1.2",
     "gatsby-plugin-react-helmet": "^3.1.16",
-    "gatsby-plugin-sharp": "^2.3.4",
+    "gatsby-plugin-sharp": "^2.3.5",
     "gatsby-plugin-styled-components": "^3.1.14",
     "gatsby-remark-copy-linked-files": "^2.1.31",
-    "gatsby-remark-images": "^3.1.34",
+    "gatsby-remark-images": "^3.1.35",
     "gatsby-remark-images-grid": "^1.0.2",
-    "gatsby-source-filesystem": "^2.1.39",
-    "gatsby-transformer-remark": "^2.6.38",
-    "gatsby-transformer-sharp": "^2.3.6",
+    "gatsby-source-filesystem": "^2.1.40",
+    "gatsby-transformer-remark": "^2.6.39",
+    "gatsby-transformer-sharp": "^2.3.7",
     "gatsby-transformer-yaml": "^2.2.18",
     "lodash": "^4.17.15",
     "node-sass": "^4.13.0",
@@ -55,7 +55,7 @@
     "react-medium-image-zoom": "^3.1.2",
     "react-slick": "^0.25.2",
     "react-uid": "^2.2.0",
-    "sharp": "^0.23.3",
+    "sharp": "^0.23.4",
     "styled-components": "^4.4.1",
     "styled-tools": "^1.7.1"
   },
@@ -65,13 +65,13 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.1",
+    "eslint": "^6.7.2",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-prettier": "^6.7.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
+    "eslint-plugin-react": "^7.17.0",
     "husky": "^3.1.0",
     "prettier": "^1.19.1"
   },
