--- src/pages/about/our-story.js
@@ -1,177 +1,0 @@
-import React from 'react';
-import { Col, Row } from 'react-flexbox-grid';
-import { Link as GatsbyLink } from 'gatsby';
-import Heading from '../../components/Heading';
-import SEO from '../../components/SEO';
-
-const OurStory = () => {
-  return (
-    <>
-      <SEO title="Our Story" />
-      <header id="story">
-        <Heading level={1} size="huge" underline className="centered">
-          Our Story
-        </Heading>
-      </header>
-      <Row className="primary-content">
-        <Col xs={12} md={12} lg={12}>
-          <p>
-            Ocius Technology Ltd (formerly Solar Sailor Holdings Ltd) is an Australian public
-            unlisted company with Research and Development facilities at the University of NSW.
-          </p>
-
-          <p>
-            In 1997, Solar Sailor won the Advanced Technology Boat Race in Canberra, beating the
-            next solar boat by 5 laps of 6km (30km) using a new patented &apos;solar sail&apos; - a
-            simple device which harnesses both solar and wind energy in a seaworthy manner.
-          </p>
-
-          <p>
-            Since then, the company has proven a solid track record of conceiving, funding, and
-            delivering difficult, innovative, high-tech, award-winning projects on time and on
-            budget.
-          </p>
-
-          <p>
-            The company has built multiple &apos;solarsailor&apos; prototypes and six high-tech,
-            award winning, hybrid electric, commercial passenger ferries to international
-            classification requirements, which together have carried tens of thousands of
-            passengers. Today, four solarsailor hybrid electric ferries operated by the Hong Kong
-            Jockey Club operate 364 days of year, 14 hours a day, from Sai Kung to Kau Sai Chau half
-            the fuel ferries.
-          </p>
-
-          <p>
-            Following an enquiry from the USA in 2007 for a &apos;platform that could go to sea
-            forever&apos;, the company began research into the development of unmanned
-            solar/wind/wave powered ocean vessels or autonomous unmanned surface vessels. Initially,
-            a 20-foot manned engineering development model (EDM) proved the &apos;speed of
-            advance&apos; in all conditions, adequate power, payload and persistence for a
-            sustainable platform go to sea for months.
-          </p>
-
-          <p>
-            In 2014, Solar Sailor changed its name to OCIUS (Latin for &apos;fleet&apos;) to reflect
-            an expansion of its patented technologies and a pivot in direction for the company.
-          </p>
-
-          <p>
-            OCIUS built multiple scale models for tank and lake testing, culminating in the building
-            of the first 10&apos; prototype called &apos;Nemo&apos; (Latin for &apos;no one&apos; ie
-            unmanned) which provided significant proof of concept of a solar, wind and wave powered
-            USV carrying a significant payload.
-          </p>
-
-          <p>
-            Based on Nemo&apos;s development, in 2015, OCIUS was awarded a Capability Technology
-            Demonstrator (CTD) from Defence Science and Technology Group. This proved a significant
-            contract and OCIUS with our partner Thales developed an 18&apos; prototype
-            &apos;Bruce&apos; for Antisubmarine Warfare (ASW), which was demonstrated to the NAVY in
-            August 2017.
-          </p>
-
-          <p>
-            In Sept 2018, Ocius was awarded its second defence contract under the Defence Innovation
-            Hub to build an intelligent “Command and Control (C2) network” of intelligent persistent
-            unmanned surface vessels.
-          </p>
-
-          <p>
-            In November 2019 &apos;Bruce&apos; participated Autonomous Warrior War games in Jervis
-            Bay. Due to extreme weather, &apos;Bruce&apos; was often the only USV operating.
-            Ocius&apos; intelligent C2 system was used to control another Navy WAMV catamaran and 2
-            DSTG UUVs.
-          </p>
-
-          <p>
-            From May - June 2019, Ocius completed 6 week endurance and intelligent networking trials
-            in a designated AMSA approved &apos;box&apos; off the NSW Coast, successfully completing
-            the DIH contract 3 months ahead of schedule.
-          </p>
-
-          <p>
-            Ocius is continuing its work with Thales, testing a new Thales thin line array deployed
-            from the Bluebottle keel winch to significant and varying depths, which is being
-            demonstrated at Pacific 2019 Expo at Darling Harbour 8-10 October.
-          </p>
-
-          <strong>Commercial Ferries</strong>
-          <p>
-            Solar Sailor built 6 commercial solar sailor vessels providing valuable experience to
-            OCIUS on the commercial maritime and technical applications of its technology as well as
-            developing an extensive network of first-class suppliers and partners.
-          </p>
-          <ul>
-            <li>
-              <strong>
-                <em>
-                  “<a href="/ck63.jpg">SolarSailor</a>
-                  ”&nbsp;
-                </em>
-              </strong>
-              <strong>(Australia).&nbsp;</strong>This vessel was completed in one year and won the
-              Australian Design Award of the Year in 2001.
-            </li>
-            <li>
-              <strong>
-                Hong Kong Jockey Club Charities Trust Project (Hong Kong).<em>&nbsp;</em>
-              </strong>
-              In July 2008, with oil at $140/ barrel, the Hong Kong Jockey Club ordered the
-              construction of four&nbsp;
-              <a href="/120116-b4-solarsails-outbound-at-sai-kung.jpg">commuter ferries</a>
-              &nbsp;to operate a service to transport players, staff and supplies from the mainland
-              at Sai Kung to the island of Kau Sai Chau where the HKJC has three 18 hole golf
-              courses. The ferries alternate two routes of five and seven mile round trips with slow
-              sectors at each end in a 20-minute cycle. Solar Sailor modelling predicted substantial
-              savings in fuel consumption compared to the previous ‘standard’ diesel ferry service,
-              as well as reductions in emissions.&nbsp; The boats in operation have exceeded this
-              since in independent tests since delivery and are still running 14 hours a day, seven
-              days a week using half the fuel of the previous ferries.
-            </li>
-            <li>
-              <strong>
-                Shanghai World Expo (China).<em>&nbsp;</em>
-              </strong>
-              After demonstration of the Sydney ferry to Suntech, a Wuxi-based solar panel company,
-              in 2009 OCIUS was awarded a contract to design, build and install a 12 -metre high
-              SolarSail with software integration, on a 250 pax VIP river cruise boat
-              vessel,&nbsp;&nbsp;‘
-              <em>
-                <GatsbyLink to="/blog/expo-ferry-rides-on-sun-and-wind-using-technology-based-on-the-evolution-of-insects-wings/">
-                  Suntech Guoshung
-                </GatsbyLink>
-              </em>
-              ’ for the World Expo in Shanghai.
-            </li>
-          </ul>
-          <p>
-            <strong>Unmanned Surface Vessels</strong>
-            <br />
-            Following an enquiry from the USA in 2007 for a ‘self-sustaining platform at sea’, OCIUS
-            began research into the development of unmanned solar/wind/wave and ballasted powered
-            ocean vessels or autonomous unmanned surface vessels. Initially, a 20 foot&nbsp;manned
-            engineering development model (EDM) proved the &apos;speed of advance&apos; in all
-            conditions, adequate power, payload and persistence for a sustainable platform able to
-            go to sea for months.
-          </p>
-          <p>
-            In 2014, Solar Sailor changed its name to OCIUS (Latin for &apos;fleet&apos;) to reflect
-            an expansion of its patented technologies including the solar sail. With Ulladulla
-            Fibreglass and Engineering, OCIUS built multiple scale models for tank and lake testing,
-            culminating in the building of the first 10&apos; prototype called &apos;Nemo&apos;
-            (Latin for &apos;no one&apos; ie unmanned) which provided significant proof of concept
-            of a solar, wind and wave powered USV carrying a significant payload.
-          </p>
-          <p>
-            Based on Nemo&apos;s development, in 2015 OCIUS was awarded a $3M Capability Technology
-            Demonstrator (CTD) from Defence Science and Technology Group. This proved a significant
-            contract and OCIUS with our partners developed an 18&apos; prototype &apos;Bruce&apos;
-            for Antisubmarine Warfare (ASW), which demonstrated to NAVY in August 2017.
-          </p>
-        </Col>
-      </Row>
-    </>
-  );
-};
-
-export default OurStory;
