--- src/pages/about/prizes.js
@@ -1,199 +1,0 @@
-import React from 'react';
-import { Link as GatsbyLink } from 'gatsby';
-import { OutboundLink } from 'gatsby-plugin-gtag';
-import { Col, Row } from 'react-flexbox-grid';
-import Heading from '../../components/Heading';
-import SEO from '../../components/SEO';
-
-const Prizes = () => {
-  return (
-    <>
-      <SEO title="Awards and Milestones" />
-      <Row>
-        <Col xs={12} md={12} lg={12}>
-          <header id="awards">
-            <Heading level={1} size="huge" underline className="centered">
-              Awards and Milestones
-            </Heading>
-          </header>
-          <Row>
-            <Col xs={12} md={6} lg={6}>
-              <p>
-                Ocius has 20 years’ experience developing and commercialising patented technology.
-              </p>
-              <p>
-                <strong>2019 April</strong> - Successful Seatrials of next-gen 18-foot Bluebottle
-                USV off NSW coast
-              </p>
-              <p>
-                <strong>2018 Nov</strong> - Success at{' '}
-                <GatsbyLink to="/blog/ocius-demonstrates-man-portable-usv-prototype-at-aw18/">
-                  Autonomous Warrior Wargames
-                </GatsbyLink>{' '}
-                with control of two USVs, two UUVs, networking to UK MAPLE C2 system and operations
-                in all conditions during three week wargames.
-              </p>
-              <p>
-                <strong>2018 Sept</strong> - Awarded{' '}
-                <GatsbyLink to="blog/defence-connect-ocius-technology-awarded-defence-innovation-hub-contract/">
-                  $1.6M DST Group Defence Innovation Hub contract
-                </GatsbyLink>{' '}
-                to demonstrate a network of intelligent persistent unmanned surface vessels
-                Australian Maritime Innovation Award: Defence SME - With a $10 000 grant
-              </p>
-              <p>
-                <strong>2017</strong> - Awarded{' '}
-                <GatsbyLink to="/blog/ocius-wins-maritime-australian-defence-innovation-sme-award/">
-                  Australian Maritime Innovation Award: Defence SME
-                </GatsbyLink>{' '}
-                - With a $10 000 grant
-              </p>
-              <p>
-                <strong>2017</strong>&nbsp;– Awarded Spitfire Memorial Defence Fellowship
-              </p>
-              <p>
-                <strong>2017</strong> – Successfully completed{' '}
-                <a href="/150914OCIUSCTDPressRelease.pdf">
-                  Capability Technology Demonstrator (CTD) Contract
-                </a>{' '}
-                for development of Anti-Submarine Unmanned Surface Vessel (ASW-USV)
-              </p>
-              <p>
-                <strong>2016</strong> – Finalist{' '}
-                <OutboundLink href="http://wind-ship.org/en/innovation_awards_2016/">
-                  Wind Propulsion Technology User Award
-                </OutboundLink>{' '}
-                &amp;{' '}
-                <GatsbyLink to="/blog/wind-propulsion-innovation-award/">
-                  Lifetime Achievement Award,{' '}
-                </GatsbyLink>
-                International Windship Association
-              </p>
-              <p>
-                <strong>2016</strong>&nbsp;–&nbsp;Robert Dane, finalist in the&nbsp;
-                <GatsbyLink to="/blog/wind-propulsion-innovation-award/">
-                  Lifetime Achievement Award,&nbsp;
-                </GatsbyLink>
-                International Windship Association
-              </p>
-              <p>
-                <strong>2016 </strong>- Awarded Spitfire Memorial Defence Scholarship
-              </p>
-              <p>
-                <strong>2015 </strong>- Build and successful testing of Bluebottle 10’ prototype
-                ‘Nemo’
-              </p>
-              <p>
-                <strong>2015</strong> –{' '}
-                <GatsbyLink to="/blog/ocius-wins-australian-technology-competition-prize/">
-                  Winner Australian Technologies Competition
-                </GatsbyLink>
-                , Best Advanced Manufacturing
-              </p>
-              <p>
-                <strong>2014</strong>&nbsp;– Sold 100 PAX<em> ‘Sydney SolarSailor’</em> ferry to
-                private buyer
-              </p>
-              <p>
-                <strong>2013</strong>&nbsp;– Energy Globe Award for Sustainability
-              </p>
-              <p>
-                <strong>2012</strong>&nbsp;– Robert Dane,&nbsp;
-                <OutboundLink href="http://www.wwf.org.au/?4023/Winners-announced-for-the-WWF-Earth-Hour-Awards">
-                  WWF Future Makers Award
-                </OutboundLink>
-              </p>
-              <p>
-                <strong>2011 </strong>–<strong>&nbsp;</strong>Delivery of the final vessel for HKJC
-                – &apos;<em>Solar Albatross</em>&apos;
-              </p>
-              <p>
-                <strong>2010</strong> –<strong>&nbsp;</strong>Winner China Seatrade Award – Dalian
-              </p>
-              <p>
-                <strong>2010 </strong>–<strong>&nbsp;</strong>Delivery of third HK ferry –{' '}
-                <em>‘Solar Eagle’</em>
-              </p>
-              <p>
-                <strong>2010</strong> –&nbsp;Delivery of second HK ferry – <em>‘Solar Birdie’</em>
-              </p>
-              <p>
-                <strong>2010 </strong>–<strong>&nbsp;</strong>Winner Environmental Technology Award
-                at the Sustainable Shipping Awards
-              </p>
-            </Col>
-            <Col xs={12} md={6} lg={6}>
-              <p>
-                <strong>2010 </strong>–&nbsp;Delivery Shanghai Vessel for World Expo – &apos;
-                <em>Suntech Guoshung</em>&apos;
-              </p>
-              <p>
-                <strong>2010 </strong>–&nbsp;Delivery first Hong Kong Jockey Club ferry –{' '}
-                <em>‘Solar Golf’</em>
-              </p>
-              <p>
-                <strong>2009&nbsp;</strong>–<strong>&nbsp;</strong>Contract Schneider Electric for
-                SS HMP190/25 supply and warranty for Hong Kong boats
-              </p>
-              <p>
-                <strong>2009&nbsp;</strong>–<strong>&nbsp;</strong>Italian Well Tech Design Award
-              </p>
-              <p>
-                <strong>2007 </strong>–&nbsp;Paper Australian Journal of Mechanical Engineers Vol 4
-                No 2. ISSN: 1448-4846
-              </p>
-              <p>
-                <strong>2007</strong>&nbsp;–<strong>&nbsp;</strong>Robert Dane&nbsp;
-                <OutboundLink href="http://www.ecocitizenaustralia.com.au/solar-sailor-dr-robert-dane/">
-                  Intel Environment Laureate USA
-                </OutboundLink>
-              </p>
-              <p>
-                <strong>2004</strong>&nbsp;–<strong>&nbsp;</strong>United Nations Awards finalist -
-                Green Ferries for Blue Highways
-              </p>
-              <p>
-                <strong>2003</strong>&nbsp;–<strong>&nbsp;</strong>Contract Captain Cook Cruises to
-                manage and operate the <em>‘Sydney SolarSailor’</em>
-              </p>
-              <p>
-                <strong>2002</strong>&nbsp;–<strong>&nbsp;</strong>Hon Bob Hawke Chairman of the
-                Company
-              </p>
-              <p>
-                <strong>2002</strong>&nbsp;–<strong>&nbsp;</strong>Robert Dane SEDA NSW Green
-                Ambassador
-              </p>
-              <p>
-                <strong>2001</strong>&nbsp;–<strong>&nbsp;</strong>Winner Australian Design Award of
-                the Year 2001
-              </p>
-              <p>
-                <strong>2000</strong>&nbsp;–<strong>&nbsp;</strong>
-                <em>‘Sydney SolarSailor’</em> launched
-              </p>
-              <p>
-                <strong>2000</strong>&nbsp;–<strong>&nbsp;</strong>Winner International Cargo
-                Handling Co-ordination Association Award
-              </p>
-              <p>
-                <strong>2000</strong>&nbsp;–<strong>&nbsp;</strong>Winner Boating Industry
-                Association of Australia Marine Awards
-              </p>
-              <p>
-                <strong>1999</strong>&nbsp;–<strong>&nbsp;</strong>Winner Far Eastern Economic
-                Review/Du Pont Asian Innovation Award
-              </p>
-              <p>
-                <strong>1997</strong>&nbsp;–<strong>&nbsp;</strong>Winner Advanced Technology Boat
-                Race Canberra – <em>‘Marjorie K’</em> prototype.
-              </p>
-            </Col>
-          </Row>
-        </Col>
-      </Row>
-    </>
-  );
-};
-
-export default Prizes;
